+++ data-export/data-export-application/src/main/kotlin/eu/tib/orkg/prototype/export/predicates/api/ExportPredicateIdToLabelUseCase.kt
@@ -1,0 +1,8 @@
+package eu.tib.orkg.prototype.export.predicates.api
+
+import java.io.Writer
+
+interface ExportPredicateIdToLabelUseCase {
+    fun export(writer: Writer)
+    fun export(path: String?)
+}