+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Predicate ids are now exported to json, along with their label.
+  (See: [!685](https://gitlab.com/TIBHannover/orkg/orkg-backend/-/merge_requests/685))
 
 ## [0.41.0] - 2023-11-10
 ### Added